Create Spin-off of "Challenge: Gradebook"
I created a database to track student grades, with their name, number grade, and what percent of activities they've completed. I first selected all of the rows, and displayed the name, number_grade, and percent_completed, which was computed by multiplying and rounding the fraction_completed column. My next goal was to create a table that shows how many students have earned which letter_grade. I created the letter_grade output by using CASE with the number_grade column, outputting 'A' for grades > 90, 'B' for grades > 80, 'C' for grades > 70, and 'F' otherwise. Then I used COUNT with GROUP BY to show the number of students with each of those grades.